extracted the imgage component
