+++ src/components/RecipePreview.js
@@ -2,16 +2,13 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 import React from 'react';
 import PropTypes from 'prop-types';
+import ReciperImgage from './RecipeImage';
 import '../css/meal.css';
 
 const Meal = ({
   src, name,
 }) => (
-
-  <div className="meal-preview shadow">
-    <img className="bg-img" src={src} alt={name} />
-    <h1 className="top-text">{name}</h1>
-  </div>
+  <ReciperImgage className="meal-preview shadow" src={src} name={name} alt={name} />
 );
 
 Meal.propTypes = {
