+++ src/css/categories.css
@@ -76,7 +76,7 @@
 }
 
 .w-100{
-    width: 100vw;
+    width: 100%;
 }
 
 .w-80{
